Update ingestion, SQL setup scripts, and tooling

+++ sql/_tmp_reset_and_prepare_dev.sql
@@ -1,6 +1,35 @@
 USE ingest_dev;
 GO
 
+-- Clean up legacy destination artifacts created under dbo schema.
+IF OBJECT_ID('dbo.dest_customers', 'U') IS NOT NULL DROP TABLE dbo.dest_customers;
+GO
+IF OBJECT_ID('dbo.dest_transactions', 'U') IS NOT NULL DROP TABLE dbo.dest_transactions;
+GO
+IF OBJECT_ID('dbo.dest_transactions_parquet', 'U') IS NOT NULL DROP TABLE dbo.dest_transactions_parquet;
+GO
+IF OBJECT_ID('dbo.dest_transactions_large', 'U') IS NOT NULL DROP TABLE dbo.dest_transactions_large;
+GO
+IF OBJECT_ID('dbo.dest_invalid_csv', 'U') IS NOT NULL DROP TABLE dbo.dest_invalid_csv;
+GO
+IF OBJECT_ID('dbo.dest_invalid_excel', 'U') IS NOT NULL DROP TABLE dbo.dest_invalid_excel;
+GO
+IF OBJECT_ID('dbo.dest_invalid_parquet', 'U') IS NOT NULL DROP TABLE dbo.dest_invalid_parquet;
+GO
+IF OBJECT_ID('dbo.sample_ingestion_target', 'U') IS NOT NULL DROP TABLE dbo.sample_ingestion_target;
+GO
+
+DECLARE @drop_sql NVARCHAR(MAX) = N'';
+SELECT @drop_sql = @drop_sql + N'DROP TABLE ' + QUOTENAME(SCHEMA_NAME(schema_id)) + N'.' + QUOTENAME(name) + N';'
+FROM sys.tables
+WHERE name LIKE '[_]tmp[_]%';
+
+IF (@drop_sql <> N'')
+BEGIN
+    EXEC sp_executesql @drop_sql;
+END
+GO
+
 IF COL_LENGTH('config.sharepoint_ingestion', 'ingestion_scope') IS NULL
 BEGIN
     ALTER TABLE config.sharepoint_ingestion ADD ingestion_scope VARCHAR(20) NOT NULL CONSTRAINT DF_sharepoint_ingestion_scope DEFAULT 'REAL';
@@ -16,6 +45,24 @@ GO
 IF COL_LENGTH('config.sharepoint_ingestion', 'is_test_data') IS NULL
 BEGIN
     ALTER TABLE config.sharepoint_ingestion ADD is_test_data BIT NOT NULL CONSTRAINT DF_sharepoint_ingestion_is_test_data DEFAULT 0;
+END
+GO
+
+IF COL_LENGTH('config.sharepoint_ingestion', 'error_notification_cc_email_address') IS NULL
+BEGIN
+    ALTER TABLE config.sharepoint_ingestion ADD error_notification_cc_email_address VARCHAR(400) NULL;
+END
+GO
+
+IF COL_LENGTH('config.sharepoint_ingestion', 'to_email_address') IS NULL
+BEGIN
+    ALTER TABLE config.sharepoint_ingestion ADD to_email_address VARCHAR(400) NULL;
+END
+GO
+
+IF COL_LENGTH('config.sharepoint_ingestion', 'cc_email_address') IS NULL
+BEGIN
+    ALTER TABLE config.sharepoint_ingestion ADD cc_email_address VARCHAR(400) NULL;
 END
 GO
 
@@ -43,22 +90,40 @@ GO
 DELETE FROM config.sharepoint_ingestion;
 GO
 
-TRUNCATE TABLE sharepoint.dest_customers;
+IF OBJECT_ID('sharepoint.dest_customers', 'U') IS NOT NULL
+    TRUNCATE TABLE sharepoint.dest_customers;
 GO
 
-TRUNCATE TABLE sharepoint.dest_transactions;
+IF OBJECT_ID('sharepoint.dest_transactions', 'U') IS NOT NULL
+    TRUNCATE TABLE sharepoint.dest_transactions;
 GO
 
-TRUNCATE TABLE sharepoint.dest_transactions_large;
+IF OBJECT_ID('sharepoint.dest_transactions_parquet', 'U') IS NOT NULL
+    TRUNCATE TABLE sharepoint.dest_transactions_parquet;
 GO
 
-TRUNCATE TABLE sharepoint.dest_invalid_csv;
+IF OBJECT_ID('sharepoint.dest_transactions_large', 'U') IS NOT NULL
+    TRUNCATE TABLE sharepoint.dest_transactions_large;
 GO
 
-TRUNCATE TABLE sharepoint.dest_invalid_excel;
+IF OBJECT_ID('sharepoint.dest_invalid_csv', 'U') IS NOT NULL
+    TRUNCATE TABLE sharepoint.dest_invalid_csv;
 GO
 
-TRUNCATE TABLE sharepoint.sample_ingestion_target;
+IF OBJECT_ID('sharepoint.dest_invalid_excel', 'U') IS NOT NULL
+    TRUNCATE TABLE sharepoint.dest_invalid_excel;
 GO
+
+IF OBJECT_ID('sharepoint.dest_invalid_parquet', 'U') IS NOT NULL
+    TRUNCATE TABLE sharepoint.dest_invalid_parquet;
+GO
+
+IF OBJECT_ID('sharepoint.sample_ingestion_target', 'U') IS NOT NULL
+    TRUNCATE TABLE sharepoint.sample_ingestion_target;
+GO
+
+
+
+
